requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
pytest==7.3.1
